Working jafu-webmvc commit
=== Building jafu-webmvc sample ===
Packaging jafu-webmvc with Maven
Unpacking jafu-webmvc-0.0.1-SNAPSHOT.jar
Performing class analysis on jafu-webmvc
Starting classpath reduction:
./compile.sh: line 49: 847079 Killed                  java
-javaagent:$CLASS_AGENT -cp $CP $MAINCLASS >> output.txt 2>&1
Compiling jafu-webmvc with GraalVM Version 20.2.0-dev (Java Version
1.8.0_252)
SUCCESS
Testing executable 'jafu-webmvc'
SUCCESS
Build memory: 5.03GB
Image build time: 50.3s
RSS memory: 29.5M
Image size: 28.0M
Startup time: 0.025 (JVM running for 0.026)

+++ spring-graal-native-samples/jafu/compile.sh
@@ -3,7 +3,7 @@
 ARTIFACT=jafu
 MAINCLASS=com.example.jafu.DemoApplication
 VERSION=0.0.1-SNAPSHOT
-FEATURE=../../../../spring-graal-native/target/spring-graal-native-0.6.0.RELEASE.jar
+FEATURE=../../../../spring-graal-native/target/spring-graal-native-0.7.0.BUILD-SNAPSHOT.jar
 
 GREEN='\033[0;32m'
 RED='\033[0;31m'
@@ -22,6 +22,11 @@ cd target/native-image
 jar -xvf ../$JAR >/dev/null 2>&1
 cp -R META-INF BOOT-INF/classes
 
+if [ ! -f "$FEATURE" ]; then
+    printf "${RED}FAILURE${NC}: $FEATURE does not exist, please build the root project before building this sample.\n"
+    exit 1
+fi
+
 # Avoids clashing substitutions from this project deps and the feature deps
 rm BOOT-INF/lib/svm-20.*.jar
 
@@ -37,6 +42,7 @@ echo "Compiling $ARTIFACT with $GRAALVM_VERSION"
   -H:Name=$ARTIFACT \
   -H:+ReportExceptionStackTraces \
   -Dspring.graal.mode=initialization-only \
+  -H:+PrintAnalysisCallTree \
   -cp $CP $MAINCLASS >> output.txt ; } 2>> output.txt
 
 if [[ -f $ARTIFACT ]]
